+++ cement/core/template.py
@@ -5,7 +5,6 @@ import sys
 import pkgutil
 import re
 import shutil
-import platform
 from abc import abstractmethod
 from ..core import exc
 from ..core.interface import Interface
@@ -159,6 +158,11 @@ class TemplateHandler(TemplateInterface, Handler):
 
         dest = fs.abspath(dest)
         src = fs.abspath(src)
+        escaped_src = src.encode('unicode-escape').decode('utf-8')
+
+        # double escape for regex matching
+        escaped_src_pattern = escaped_src.encode('unicode-escape')
+        escaped_src_pattern = escaped_src_pattern.decode('utf-8')
 
         if exclude is None:
             exclude = []
@@ -176,6 +180,8 @@ class TemplateHandler(TemplateInterface, Handler):
 
         # here's the fun
         for cur_dir, sub_dirs, files in os.walk(src):
+            escaped_cur_dir = cur_dir.encode('unicode-escape').decode('utf-8')
+
             if cur_dir == '.':
                 continue    # pragma: nocover
             elif cur_dir == src:
@@ -190,24 +196,33 @@ class TemplateHandler(TemplateInterface, Handler):
                 LOG.debug(
                     'not rendering excluded directory as template: ' +
                     '%s' % cur_dir)
-                cur_dir_stub = re.sub(src, '', cur_dir)
+
+                cur_dir_stub = re.sub(escaped_src_pattern,
+                                      '',
+                                      escaped_cur_dir)
                 cur_dir_stub = cur_dir_stub.lstrip('/')
+                cur_dir_stub = cur_dir_stub.lstrip('\\\\')
                 cur_dir_stub = cur_dir_stub.lstrip('\\')
                 cur_dir_dest = os.path.join(dest, cur_dir_stub)
             else:
                 # render the cur dir
                 LOG.debug(
                     'rendering directory as template: %s' % cur_dir)
-                cur_dir_stub = re.sub(src,
-                                      '',
-                                      self.render(cur_dir, data))
 
+                cur_dir_stub = re.sub(escaped_src_pattern,
+                                      '',
+                                      escaped_cur_dir)
+                cur_dir_stub = self.render(cur_dir_stub, data)
                 cur_dir_stub = cur_dir_stub.lstrip('/')
+                cur_dir_stub = cur_dir_stub.lstrip('\\\\')
                 cur_dir_stub = cur_dir_stub.lstrip('\\')
                 cur_dir_dest = os.path.join(dest, cur_dir_stub)
 
             # render sub-dirs
             for sub_dir in sub_dirs:
+                escaped_sub_dir = sub_dir.encode('unicode-escape')
+                escaped_sub_dir = escaped_sub_dir.decode('utf-8')
+
                 full_path = os.path.join(cur_dir, sub_dir)
 
                 if self._match_patterns(full_path, ignore_patterns):
@@ -224,13 +239,9 @@ class TemplateHandler(TemplateInterface, Handler):
                     LOG.debug(
                         'rendering sub-directory as template: %s' % full_path)
 
-                    if platform.system().lower() in ['windows']:
-                        src = src.encode('unicode-escape')  # pragma: nocover
-                        src = str(src)                      # pragma: nocover
-
-                    new_sub_dir = re.sub(src,
+                    new_sub_dir = re.sub(escaped_src_pattern,
                                          '',
-                                         self.render(sub_dir, data))
+                                         self.render(escaped_sub_dir, data))
                     sub_dir_dest = os.path.join(cur_dir_dest, new_sub_dir)
 
                 if not os.path.exists(sub_dir_dest):
@@ -238,12 +249,9 @@ class TemplateHandler(TemplateInterface, Handler):
                     os.makedirs(sub_dir_dest)
 
             for _file in files:
-                # windows paths (if condition to avoid unknown breakage)
-                if platform.system().lower() in ['windows']:
-                    src = src.encode('unicode-escape')      # pragma: nocover
-                    src = str(src)                          # pragma: nocover
+                _rendered = self.render(_file, data)
+                new_file = re.sub(escaped_src_pattern, '', _rendered)
 
-                new_file = re.sub(src, '', self.render(_file, data))
                 _file = fs.abspath(os.path.join(cur_dir, _file))
                 _file_dest = fs.abspath(os.path.join(cur_dir_dest, new_file))